+++ ism7mqtt/run.sh
@@ -6,11 +6,14 @@ export DEBUG_LOGGING=$(bashio::config 'debug_logging')
 
 
 export ISM7_MQTTHOST="$(bashio::config 'mqtt_host')"
+export ISM7_MQTTPORT="$(bashio::config 'mqtt_port')"
 export ISM7_MQTTUSERNAME="$(bashio::config 'mqtt_user')"
 export ISM7_MQTTPASSWORD="$(bashio::config 'mqtt_password')"
 
+
 if [[ "$ISM7_MQTTHOST" == "null" ]]; then
     export ISM7_MQTTHOST="$(bashio::services mqtt 'host')"
+    export ISM7_MQTTPORT="$(bashio::services mqtt 'port')"
     export ISM7_MQTTUSERNAME="$(bashio::services mqtt 'username')"
     export ISM7_MQTTPASSWORD="$(bashio::services mqtt 'password')"
     echo "Reading config from MQTT broker add-on: $ISM7_MQTTHOST/$ISM7_MQTTUSERNAME"